Enable new rubocop cops
Since we update to new rubocop versions we never activated new cops.
Rubocop by default will not apply newly introduced rules.
This PR enables the latest cops and add the autocorrections of rubocop.
Additionally a rubocop_todo.yml file will ignore present failures so
that they apply for new code and one has a comprehensive list of
findings on can one by one fix in the future.

+++ spec/support/bootstrap/spec_bootstrap.rb
@@ -17,9 +17,7 @@ module VCAP::CloudController
 
       log_filename = File.join(Paths::ARTIFACTS, 'spec.log')
       max_log_file_size_in_bytes = 100_000_000
-      if File.exist?(log_filename) && File.size(log_filename) > max_log_file_size_in_bytes
-        FileUtils.rm_f(log_filename)
-      end
+      FileUtils.rm_f(log_filename) if File.exist?(log_filename) && File.size(log_filename) > max_log_file_size_in_bytes
       logger = ActiveSupport::Logger.new(File.join(Paths::ARTIFACTS, 'telemetry_spec.log'))
       TelemetryLogger.init(logger)
